R09/T02/feature/side_bar | Componentização dos elementos - criei a pasta "side_bar" em global/widgets - dentro da pasta side_bar, há uma pasta "components_side_bar" e dois arquivos: o primeiro contem as classes referentes aos botões; - no segundo, os métodos são referenciados passando o Text dos botões
