+++ lib/app/modules/restaurant_home/restaurant_home_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter/material.dart';
+import 'package:pscomidas/app/global/widgets/side_bar/components_side_bar/menu_side_bar.dart';
+import 'package:pscomidas/app/modules/home/schemas.dart';
 import 'package:pscomidas/app/modules/restaurant_home/restaurant_home_store.dart';
-import 'package:pscomidas/app/modules/restaurant_home/side_bar_components/menu_side_bar.dart';
 
 class RestaurantHomePage extends StatefulWidget {
   final String title;
@@ -23,13 +24,24 @@ class RestaurantHomePageState extends State<RestaurantHomePage> {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             //side bar
-            Container(
+            SizedBox(
               width: screen.width * 0.2,
               child: const SideBar(),
             ),
             //main body
             Expanded(
-              child: Container(),
+              // ignore: avoid_unnecessary_containers
+              child: Container(
+                child: const Text(
+                  "Produtos",
+                  style: TextStyle(
+                    color: secondaryCollor,
+                    fontFamily: "Nunito",
+                    fontSize: 40,
+                  ),
+                  textAlign: TextAlign.center,
+                ),
+              ),
             ),
           ],
         ),
